use environment variables for tokenurl, tokeninfo url, cert_path and scopes. Send them to J4J_Orchestrator in the header
