+++ j4j_authenticator/__init__.py
@@ -1,2 +1,2 @@
-from .base_auth import BaseAuthenticator
+from .j4j_authenticator import BaseAuthenticator
 from .j4j_logout import J4J_LogoutHandler
