+++ j4j_spawner/utils.py
@@ -1,4 +1,5 @@
 import json
+import os
 import requests
 
 from concurrent.futures import ThreadPoolExecutor
@@ -6,7 +7,7 @@ from contextlib import closing
 
 from .file_loads import get_token
 
-def create_spawn_header(uuidcode, expire, refreshtoken, jhubtoken, accesstoken, account, project, servername, escapedusername, orchestrator_token_path):
+def create_spawn_header(uuidcode, expire, refreshtoken, jhubtoken, accesstoken, account, project, servername, escapedusername, orchestrator_token_path, login_handler):
     spawn_header = {
         "uuidcode": uuidcode,
         "Intern-Authorization": get_token(orchestrator_token_path),
@@ -20,6 +21,21 @@ def create_spawn_header(uuidcode, expire, refreshtoken, jhubtoken, accesstoken, 
         "servername": servername,
         "escapedusername": escapedusername
         }
+    if login_handler == 'jscldap':
+        spawn_header['tokenurl'] = os.environ.get('JSCLDAP_TOKEN_URL', '')
+        spawn_header['tokeninfourl'] = os.environ.get('JSCLDAP_TOKENINFO_URL', '')
+        spawn_header['certpath'] = os.environ.get('JSCLDAP_CERT_PATH', '')
+        spawn_header['scope'] = os.environ.get('JSCLDAP_SCOPE', '')
+    elif login_handler == 'jscworkshop':
+        spawn_header['tokenurl'] = os.environ.get('JSCWORKSHOP_TOKEN_URL', '')
+        spawn_header['tokeninfourl'] = os.environ.get('JSCWORKSHOP_TOKENINFO_URL', '')
+        spawn_header['certpath'] = os.environ.get('JSCWORKSHOP_CERT_PATH', '')
+        spawn_header['scope'] = os.environ.get('JSCWORKSHOP_SCOPE', '')
+    else:
+        spawn_header['tokenurl'] = ''
+        spawn_header['tokeninfourl'] = ''
+        spawn_header['certpath'] = ''
+        spawn_header['scope'] = ''
     return spawn_header
 
 def create_spawn_data(servername, Environment, partition, reservation, Resources, system, Checkboxes):
